requirements.txt
gspread==6.0.0
google-auth==2.23.0
pandas==2.1.1
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1